detect/http_header: move tests into tests/
